Mana con float

+++ src/Managers/SpellManager.cpp
@@ -86,7 +86,7 @@ bool SpellManager::StartHechizo(int num, Player* p){
 		if(hechizos[num].find(p) != hechizos[num].end()){	// Comprobamos que la clave este
 			Hechizo* h = hechizos[num][p];			// Cargamos el hechizo en una variables
 			if(h!=NULL){							// Comprobamos si realmente existe
-				int mana = p->GetMP();
+				float mana = p->GetMP();
 				if(h->CheckMP(mana)){
 					h->EmpezarCast();
 					return true;
